# Optimize page buffer allocations